hotbar align to pixels and java 17

+++ src/main/java/io/github/niterux/niterucks/mixin/ui/GameGuiMixin.java
@@ -52,7 +52,7 @@ public class GameGuiMixin extends GuiElement {
 	}
 
 	@Unique
-	double hotbarHeight;
+	int hotbarHeight;
 	@Shadow
 	private Minecraft minecraft;
